requirements.txt
duckdb
pandas
langgraph
langchain
langchain-community
python-dotenv
rich
great-expectations
